feat(tests): factor out base open id connect test class
This allows better type checking for related tests and fixes some
possible issues in tests.
