+++ social_core/tests/backends/test_elixir.py
@@ -1,8 +1,8 @@
-from .oauth import BaseAuthUrlTestMixin, OAuth2Test
-from .test_open_id_connect import OpenIdConnectTestMixin
+from .oauth import BaseAuthUrlTestMixin
+from .open_id_connect import OpenIdConnectTest
 
 
-class ElixirOpenIdConnectTest(OpenIdConnectTestMixin, OAuth2Test, BaseAuthUrlTestMixin):
+class ElixirOpenIdConnectTest(OpenIdConnectTest, BaseAuthUrlTestMixin):
     backend_path = "social_core.backends.elixir.ElixirOpenIdConnect"
     issuer = "https://login.elixir-czech.org/oidc/"
     openid_config_body = """
