+++ social_core/tests/backends/test_fence.py
@@ -1,10 +1,10 @@
 import json
 
-from .oauth import BaseAuthUrlTestMixin, OAuth2Test
-from .test_open_id_connect import OpenIdConnectTestMixin
+from .oauth import BaseAuthUrlTestMixin
+from .open_id_connect import OpenIdConnectTest
 
 
-class FenceOpenIdConnectTest(OpenIdConnectTestMixin, OAuth2Test, BaseAuthUrlTestMixin):
+class FenceOpenIdConnectTest(OpenIdConnectTest, BaseAuthUrlTestMixin):
     backend_path = "social_core.backends.fence.Fence"
     issuer = "https://nci-crdc.datacommons.io/"
     openid_config_body = json.dumps(
