+++ social_core/tests/backends/test_lifescience.py
@@ -1,12 +1,10 @@
 import json
 
-from .oauth import BaseAuthUrlTestMixin, OAuth2Test
-from .test_open_id_connect import OpenIdConnectTestMixin
+from .oauth import BaseAuthUrlTestMixin
+from .open_id_connect import OpenIdConnectTest
 
 
-class LifeScienceOpenIdConnectTest(
-    OpenIdConnectTestMixin, OAuth2Test, BaseAuthUrlTestMixin
-):
+class LifeScienceOpenIdConnectTest(OpenIdConnectTest, BaseAuthUrlTestMixin):
     backend_path = "social_core.backends.lifescience.LifeScienceOpenIdConnect"
     issuer = "https://login.aai.lifescience-ri.eu/oidc/"
     user_data_url = "https://login.aai.lifescience-ri.eu/oidc/userinfo"
