+++ social_core/tests/backends/test_linkedin.py
@@ -1,14 +1,10 @@
-# pyright: reportAttributeAccessIssue=false
-
 import json
 
 from .oauth import BaseAuthUrlTestMixin, OAuth2Test
-from .test_open_id_connect import OpenIdConnectTestMixin
+from .open_id_connect import OpenIdConnectTest
 
 
-class LinkedinOpenIdConnectTest(
-    OpenIdConnectTestMixin, OAuth2Test, BaseAuthUrlTestMixin
-):
+class LinkedinOpenIdConnectTest(OpenIdConnectTest, BaseAuthUrlTestMixin):
     backend_path = "social_core.backends.linkedin.LinkedinOpenIdConnect"
     user_data_url = "https://api.linkedin.com/v2/userinfo"
     issuer = "https://www.linkedin.com"
@@ -66,10 +62,10 @@ class BaseLinkedinTest:
     )
 
     def test_login(self):
-        self.do_login()
+        self.do_login()  # type: ignore[attr-defined]
 
     def test_partial_pipeline(self):
-        self.do_partial_pipeline()
+        self.do_partial_pipeline()  # type: ignore[attr-defined]
 
 
 class LinkedinOAuth2Test(BaseLinkedinTest, OAuth2Test):
